% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.confint.glsm.R
\name{print.confint.glsm}
\alias{print.confint.glsm}
\title{Print method for confidence intervals of a GLSM}
\usage{
\method{print}{confint.glsm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{confint.glsm}.}

\item{...}{Further arguments (unused).}
}
\value{
Prints output to the console.
}
\description{
Prints the confidence intervals and odds ratios computed by \code{\link[=confint.glsm]{confint.glsm()}}.
}
