% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.glsm.R
\name{print.summary.glsm}
\alias{print.summary.glsm}
\title{Print method for summary.glsm objects}
\usage{
\method{print}{summary.glsm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.glsm}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Prints the summary tables to the console. Invisibly returns \code{NULL}.
}
\description{
Nicely prints the output of \code{\link[=summary.glsm]{summary.glsm()}}, including coefficients and deviance comparisons.
}
\seealso{
\code{\link[=summary.glsm]{summary.glsm()}}, \code{\link[=glsm]{glsm()}}
}
