% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glsm.R
\name{glsm}
\alias{glsm}
\title{Generalized Logistic Saturated Model (GLSM)}
\usage{
glsm(formula, data, ref = NaN)
}
\arguments{
\item{formula}{A formula specifying the dependent and independent variables, e.g., \code{y ~ x1 + x2}.}

\item{data}{A \code{data.frame} containing the variables used in the formula.}

\item{ref}{A character string indicating the reference category of the dependent variable. If \code{NaN}, the first level is used by default.}
}
\value{
An object of class \code{glsm} containing:
\describe{
\item{coefficients}{The estimated coefficients.}
\item{Std.Error}{Standard errors of the coefficients.}
\item{ExpB}{Exponentiated coefficients (odds ratios).}
\item{Log.Lik}{Log-likelihoods for the different models.}
\item{Deviance}{Deviances and tests for model comparisons.}
\item{Odds}{Fitted odds and odds ratios.}
\item{Probabilities}{Fitted category probabilities.}
\item{call}{The original function call.}
}
}
\description{
Fits a multinomial logistic regression model with a saturated structure.
The function estimates the coefficients using an Iteratively Reweighted Least Squares (IRLS) algorithm,
calculates log-likelihoods for the null, complete, and saturated models, and returns detailed outputs
for model comparison, predicted probabilities, odds, and odds ratios.
}
\details{
This implementation works for categorical outcomes with 3 or more levels.
For binary logistic regression, use \code{lsm()} instead.
The function automatically computes:
\itemize{
\item Log-likelihoods for null, complete, logit, and saturated models.
\item Deviance and p-values for model comparisons.
\item Coefficients, standard errors, Wald statistics, and odds ratios.
}
}
\examples{
\dontrun{
data(mydata)
model <- glsm(prog ~ ses + write + read, data = mydata, ref = "academic")
summary(model)
confint(model, parm = "write:vocation")
predict(model, type = "response")
plot(model)
}

}
